split patch/export.go into two parts.
