+++ cmd/registry/patch/export.go
@@ -27,145 +27,6 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-func buildAPI(ctx context.Context, client *gapic.RegistryClient, message *rpc.Api) (*API, error) {
-	apiName, err := names.ParseApi(message.Name)
-	if err != nil {
-		return nil, err
-	}
-	api := &API{
-		Header: Header{
-			APIVersion: REGISTRY_V1,
-			Kind:       "API",
-			Metadata: Metadata{
-				Name:        apiName.ApiID,
-				Labels:      message.Labels,
-				Annotations: message.Annotations,
-			},
-		},
-	}
-	api.Spec.DisplayName = message.DisplayName
-	api.Spec.Description = message.Description
-	api.Spec.Availability = message.Availability
-	api.Spec.RecommendedVersion = message.RecommendedVersion
-	api.Spec.RecommendedDeployment = message.RecommendedDeployment
-	core.ListVersions(ctx, client, apiName.Version(""), "", func(message *rpc.ApiVersion) {
-		version, err2 := buildAPIVersion(ctx, client, message)
-		// unset these because they can be inferred
-		version.APIVersion = ""
-		version.Kind = ""
-		if err2 == nil {
-			api.Spec.APIVersions = append(api.Spec.APIVersions, version)
-		} else {
-			err = err2
-		}
-	})
-	core.ListDeployments(ctx, client, apiName.Deployment(""), "", func(message *rpc.ApiDeployment) {
-		deployment, err2 := buildAPIDeployment(ctx, client, message)
-		// unset these because they can be inferred
-		deployment.APIVersion = ""
-		deployment.Kind = ""
-		if err2 == nil {
-			api.Spec.APIDeployments = append(api.Spec.APIDeployments, deployment)
-		} else {
-			err = err2
-		}
-	})
-	return api, err
-}
-
-func buildAPIVersion(ctx context.Context, client *gapic.RegistryClient, message *rpc.ApiVersion) (*APIVersion, error) {
-	versionName, err := names.ParseVersion(message.Name)
-	if err != nil {
-		return nil, err
-	}
-	version := &APIVersion{
-		Header: Header{
-			APIVersion: REGISTRY_V1,
-			Kind:       "APIVersion",
-			Metadata: Metadata{
-				Name:        versionName.VersionID,
-				Labels:      message.Labels,
-				Annotations: message.Annotations,
-			},
-		},
-	}
-	version.Spec.DisplayName = message.DisplayName
-	version.Spec.Description = message.Description
-	core.ListSpecs(ctx, client, versionName.Spec(""), "", func(message *rpc.ApiSpec) {
-		spec, err2 := buildAPISpec(ctx, client, message)
-		// unset these because they can be inferred
-		spec.APIVersion = ""
-		spec.Kind = ""
-		if err2 == nil {
-			version.Spec.APISpecs = append(version.Spec.APISpecs, spec)
-		} else {
-			err = err2
-		}
-	})
-	return version, err
-}
-
-func buildAPISpec(ctx context.Context, client *gapic.RegistryClient, message *rpc.ApiSpec) (*APISpec, error) {
-	specName, err := names.ParseSpec(message.Name)
-	if err != nil {
-		return nil, err
-	}
-	spec := &APISpec{
-		Header: Header{
-			APIVersion: REGISTRY_V1,
-			Kind:       "APISpec",
-			Metadata: Metadata{
-				Name:        specName.SpecID,
-				Labels:      message.Labels,
-				Annotations: message.Annotations,
-			},
-		},
-	}
-	spec.Spec.FileName = message.Filename
-	spec.Spec.Description = message.Description
-	spec.Spec.MimeType = message.MimeType
-	spec.Spec.SourceURI = message.SourceUri
-	return spec, nil
-}
-
-func buildAPIDeployment(ctx context.Context, client *gapic.RegistryClient, message *rpc.ApiDeployment) (*APIDeployment, error) {
-	deploymentName, err := names.ParseDeployment(message.Name)
-	if err != nil {
-		return nil, err
-	}
-	deployment := &APIDeployment{
-		Header: Header{
-			APIVersion: REGISTRY_V1,
-			Kind:       "APIDeployment",
-			Metadata: Metadata{
-				Name:        deploymentName.DeploymentID,
-				Labels:      message.Labels,
-				Annotations: message.Annotations,
-			},
-		},
-	}
-	deployment.Spec.DisplayName = message.DisplayName
-	deployment.Spec.Description = message.Description
-	return deployment, nil
-}
-
-func buildArtifact(message *rpc.Artifact) (*Artifact, error) {
-	artifactName, err := names.ParseArtifact(message.Name)
-	if err != nil {
-		return nil, err
-	}
-	artifact := &Artifact{
-		Header: Header{
-			APIVersion: REGISTRY_V1,
-			Kind:       "Artifact",
-			Metadata: Metadata{
-				Name: artifactName.ArtifactID(),
-			},
-		},
-	}
-	return artifact, nil
-}
-
 // ExportProject writes a project as a YAML file.
 func ExportProject(ctx context.Context, client *gapic.RegistryClient, adminClient *gapic.AdminClient, message *rpc.Project) {
 	projectName, err := names.ParseProject(message.Name)
